Completed more of the theory work
Today I completed most of testing and debugging and all the development that I can do. I will be documenting commits both in VSCode and GitHub Desktop from now on. The aim for today was to complete the structure charts, however, I wish to take a break due to completing a substantial amount of work compared to the amount of progress that I have made so far. In the future, I shall finish the structure chart for the program and work on the 'Integration' section, which will make me up to scratch for completion on Monday 7/4.