Completed Integration and Development
Today I completed the Development and Integration parts of the theory. I also added docstrings to the main functions, as well as completed the requirements.txt file. This means that I only have the Installation and Maintenance parts to be completed, which are really short. I likely will not complete the task by tomorrow as mentioned in my previous commits, however, maybe I will be done by Wednesday.